[Benchmark] Move problem file to the test_cases subfolder
