+++ benchmark/catalyst_benchmark/measurements.py
@@ -117,7 +117,11 @@ def measure_compile_catalyst(a: ParsedArguments) -> BenchmarkResult:
 
     p: Problem  # pylint: disable=used-before-assignment
     if a.problem == "grover":
-        from catalyst_benchmark.grover_catalyst import ProblemC as Problem, qcompile, workflow
+        from catalyst_benchmark.test_cases.grover_catalyst import (
+            ProblemC as Problem,
+            qcompile,
+            workflow,
+        )
 
         p = Problem(
             qml.device("lightning.qubit", wires=a.nqubits),
@@ -125,7 +129,11 @@ def measure_compile_catalyst(a: ParsedArguments) -> BenchmarkResult:
             expansion_strategy="device",
         )
     elif a.problem == "chemvqe":
-        from catalyst_benchmark.chemvqe_catalyst import ProblemCVQE as Problem, qcompile, workflow
+        from catalyst_benchmark.test_cases.chemvqe_catalyst import (
+            ProblemCVQE as Problem,
+            qcompile,
+            workflow,
+        )
 
         p = Problem(
             qml.device("lightning.qubit", wires=a.nqubits),
@@ -133,7 +141,7 @@ def measure_compile_catalyst(a: ParsedArguments) -> BenchmarkResult:
             expansion_strategy="device",
         )
     elif a.problem == "chemvqe-hybrid":
-        from catalyst_benchmark.chemvqe_catalyst import (
+        from catalyst_benchmark.test_cases.chemvqe_catalyst import (
             ProblemCVQE as Problem,
             qcompile_hybrid as qcompile,
             workflow_hybrid as workflow,
@@ -145,7 +153,11 @@ def measure_compile_catalyst(a: ParsedArguments) -> BenchmarkResult:
             expansion_strategy="device",
         )
     elif a.problem == "qft":
-        from catalyst_benchmark.qft_catalyst import ProblemC as Problem, qcompile, workflow
+        from catalyst_benchmark.test_cases.qft_catalyst import (
+            ProblemC as Problem,
+            qcompile,
+            workflow,
+        )
 
         p = Problem(
             qml.device("lightning.qubit", wires=a.nqubits), a.nlayers, expansion_strategy="device"
@@ -191,16 +203,24 @@ def measure_runtime_catalyst(a: ParsedArguments) -> BenchmarkResult:
 
     p: Problem  # pylint: disable=used-before-assignment
     if a.problem == "grover":
-        from catalyst_benchmark.grover_catalyst import ProblemC as Problem, qcompile, workflow
+        from catalyst_benchmark.test_cases.grover_catalyst import (
+            ProblemC as Problem,
+            qcompile,
+            workflow,
+        )
 
         p = Problem(qml.device("lightning.qubit", wires=a.nqubits), a.nlayers)
 
     elif a.problem == "chemvqe":
-        from catalyst_benchmark.chemvqe_catalyst import ProblemCVQE as Problem, qcompile, workflow
+        from catalyst_benchmark.test_cases.chemvqe_catalyst import (
+            ProblemCVQE as Problem,
+            qcompile,
+            workflow,
+        )
 
         p = Problem(qml.device("lightning.qubit", wires=a.nqubits), diff_method=a.vqe_diff_method)
     elif a.problem == "chemvqe-hybrid":
-        from catalyst_benchmark.chemvqe_catalyst import (
+        from catalyst_benchmark.test_cases.chemvqe_catalyst import (
             ProblemCVQE as Problem,
             qcompile_hybrid as qcompile,
             workflow_hybrid as workflow,
@@ -208,7 +228,11 @@ def measure_runtime_catalyst(a: ParsedArguments) -> BenchmarkResult:
 
         p = Problem(qml.device("lightning.qubit", wires=a.nqubits), diff_method=a.vqe_diff_method)
     elif a.problem == "qft":
-        from catalyst_benchmark.qft_catalyst import ProblemC as Problem, qcompile, workflow
+        from catalyst_benchmark.test_cases.qft_catalyst import (
+            ProblemC as Problem,
+            qcompile,
+            workflow,
+        )
 
         p = Problem(
             qml.device("lightning.qubit", wires=a.nqubits),
@@ -257,7 +281,7 @@ def measure_compile_pennylanejax(a: ParsedArguments) -> BenchmarkResult:
 
     p: Problem  # pylint: disable=used-before-assignment
     if a.problem == "grover":
-        from catalyst_benchmark.grover_pennylane import (
+        from catalyst_benchmark.test_cases.grover_pennylane import (
             ProblemPL as Problem,
             qcompile,
             workflow,
@@ -272,7 +296,7 @@ def measure_compile_pennylanejax(a: ParsedArguments) -> BenchmarkResult:
         )
 
     elif a.problem == "chemvqe":
-        from catalyst_benchmark.chemvqe_pennylane import (
+        from catalyst_benchmark.test_cases.chemvqe_pennylane import (
             ProblemCVQE as Problem,
             qcompile,
             workflow,
@@ -287,7 +311,7 @@ def measure_compile_pennylanejax(a: ParsedArguments) -> BenchmarkResult:
             expansion_strategy="device",
         )
     elif a.problem == "chemvqe-hybrid":
-        from catalyst_benchmark.chemvqe_pennylane import (
+        from catalyst_benchmark.test_cases.chemvqe_pennylane import (
             ProblemCVQE as Problem,
             qcompile_hybrid as qcompile,
             workflow_hybrid as workflow,
@@ -302,7 +326,12 @@ def measure_compile_pennylanejax(a: ParsedArguments) -> BenchmarkResult:
             expansion_strategy="device",
         )
     elif a.problem == "qft":
-        from catalyst_benchmark.qft_pennylane import ProblemPL as Problem, qcompile, workflow, size
+        from catalyst_benchmark.test_cases.qft_pennylane import (
+            ProblemPL as Problem,
+            qcompile,
+            workflow,
+            size,
+        )
 
         p = Problem(
             qml.device(device, wires=a.nqubits),
@@ -351,7 +380,7 @@ def measure_runtime_pennylanejax(a: ParsedArguments) -> BenchmarkResult:
 
     p: Problem  # pylint: disable=used-before-assignment
     if a.problem == "grover":
-        from catalyst_benchmark.grover_pennylane import (
+        from catalyst_benchmark.test_cases.grover_pennylane import (
             ProblemPL as Problem,
             qcompile,
             workflow,
@@ -361,7 +390,7 @@ def measure_runtime_pennylanejax(a: ParsedArguments) -> BenchmarkResult:
         p = Problem(qml.device(device, wires=a.nqubits), a.nlayers, interface=interface)
 
     elif a.problem == "chemvqe":
-        from catalyst_benchmark.chemvqe_pennylane import (
+        from catalyst_benchmark.test_cases.chemvqe_pennylane import (
             ProblemCVQE as Problem,
             qcompile,
             workflow,
@@ -375,7 +404,7 @@ def measure_runtime_pennylanejax(a: ParsedArguments) -> BenchmarkResult:
             diff_method=a.vqe_diff_method,
         )
     elif a.problem == "chemvqe-hybrid":
-        from catalyst_benchmark.chemvqe_pennylane import (
+        from catalyst_benchmark.test_cases.chemvqe_pennylane import (
             ProblemCVQE as Problem,
             qcompile_hybrid as qcompile,
             workflow_hybrid as workflow,
@@ -389,7 +418,12 @@ def measure_runtime_pennylanejax(a: ParsedArguments) -> BenchmarkResult:
             diff_method=a.vqe_diff_method,
         )
     elif a.problem == "qft":
-        from catalyst_benchmark.qft_pennylane import ProblemPL as Problem, qcompile, workflow, size
+        from catalyst_benchmark.test_cases.qft_pennylane import (
+            ProblemPL as Problem,
+            qcompile,
+            workflow,
+            size,
+        )
 
         p = Problem(qml.device(device, wires=a.nqubits), interface=interface, nlayers=a.nlayers)
 
@@ -430,7 +464,7 @@ def measure_compile_pennylane(a: ParsedArguments) -> BenchmarkResult:
 
     p: Problem  # pylint: disable=used-before-assignment
     if a.problem == "grover":
-        from catalyst_benchmark.grover_pennylane import (
+        from catalyst_benchmark.test_cases.grover_pennylane import (
             ProblemPL as Problem,
             qcompile,
             workflow,
@@ -444,7 +478,7 @@ def measure_compile_pennylane(a: ParsedArguments) -> BenchmarkResult:
             expansion_strategy="device",
         )
     elif a.problem == "chemvqe":
-        from catalyst_benchmark.chemvqe_pennylane import (
+        from catalyst_benchmark.test_cases.chemvqe_pennylane import (
             ProblemCVQE as Problem,
             qcompile,
             workflow,
@@ -458,7 +492,7 @@ def measure_compile_pennylane(a: ParsedArguments) -> BenchmarkResult:
             expansion_strategy="device",
         )
     elif a.problem == "chemvqe-hybrid":
-        from catalyst_benchmark.chemvqe_pennylane import (
+        from catalyst_benchmark.test_cases.chemvqe_pennylane import (
             ProblemCVQE as Problem,
             qcompile_hybrid as qcompile,
             workflow_hybrid as workflow,
@@ -472,7 +506,12 @@ def measure_compile_pennylane(a: ParsedArguments) -> BenchmarkResult:
             expansion_strategy="device",
         )
     elif a.problem == "qft":
-        from catalyst_benchmark.qft_pennylane import ProblemPL as Problem, qcompile, workflow, size
+        from catalyst_benchmark.test_cases.qft_pennylane import (
+            ProblemPL as Problem,
+            qcompile,
+            workflow,
+            size,
+        )
 
         p = Problem(
             qml.device(device, wires=a.nqubits),
@@ -509,7 +548,7 @@ def measure_runtime_pennylane(a: ParsedArguments) -> BenchmarkResult:
 
     p: Problem  # pylint: disable=used-before-assignment
     if a.problem == "grover":
-        from catalyst_benchmark.grover_pennylane import (
+        from catalyst_benchmark.test_cases.grover_pennylane import (
             ProblemPL as Problem,
             workflow,
             qcompile,
@@ -518,7 +557,7 @@ def measure_runtime_pennylane(a: ParsedArguments) -> BenchmarkResult:
 
         p = Problem(qml.device(device, wires=a.nqubits), a.nlayers)
     elif a.problem == "chemvqe":
-        from catalyst_benchmark.chemvqe_pennylane import (
+        from catalyst_benchmark.test_cases.chemvqe_pennylane import (
             ProblemCVQE as Problem,
             qcompile,
             workflow,
@@ -531,7 +570,7 @@ def measure_runtime_pennylane(a: ParsedArguments) -> BenchmarkResult:
             diff_method=a.vqe_diff_method,
         )
     elif a.problem == "chemvqe-hybrid":
-        from catalyst_benchmark.chemvqe_pennylane import (
+        from catalyst_benchmark.test_cases.chemvqe_pennylane import (
             ProblemCVQE as Problem,
             qcompile_hybrid as qcompile,
             workflow_hybrid as workflow,
@@ -544,7 +583,12 @@ def measure_runtime_pennylane(a: ParsedArguments) -> BenchmarkResult:
             diff_method=a.vqe_diff_method,
         )
     elif a.problem == "qft":
-        from catalyst_benchmark.qft_pennylane import ProblemPL as Problem, qcompile, workflow, size
+        from catalyst_benchmark.test_cases.qft_pennylane import (
+            ProblemPL as Problem,
+            qcompile,
+            workflow,
+            size,
+        )
 
         p = Problem(
             qml.device(device, wires=a.nqubits),
