+++ benchmark/catalyst_benchmark/test_cases/grover_pennylane.py
@@ -20,7 +20,7 @@ from dataclasses import dataclass
 
 import pennylane as qml
 import pennylane.numpy as pnp
-from .types import Problem
+from catalyst_benchmark.types import Problem
 
 
 def clause_nqubits(clause_list) -> int:
